+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 - Fixed invalid code completion for keywords using meta characters like `*` or `/` with a variable definition before the keyword
 - Freeze of UI when starting debug mode
 - Fixed detection of locations a breakpoint in robot file is settable or not
+- Fixed log warn messages indicating about more than one serialization of module configuration
 
 ### Changed